# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/szemelyek": {
      "get": {
        "description": "Lekéri az összes személyt, szűrhető a megadott paraméterek alapján.",
        "operationId": "SzemelyekController_findAll",
        "parameters": [
          {
            "name": "szemely_email",
            "required": false,
            "in": "query",
            "description": "Szűrő az email címre",
            "schema": {}
          },
          {
            "name": "szemely_keresztnev",
            "required": false,
            "in": "query",
            "description": "Szűrő a keresztnévre",
            "schema": {}
          },
          {
            "name": "szemely_vezeteknev",
            "required": false,
            "in": "query",
            "description": "Szűrő a vezetéknévre",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Személyek sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Szemely"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin vagy user jogosultság"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Összes személy lekérdezése",
        "tags": [
          "Személyek"
        ]
      },
      "post": {
        "description": "Új személy regisztrálása admin jogosultsággal.",
        "operationId": "SzemelyekController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSzemelyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Személy sikeresen létrehozva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Új személy létrehozása",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/{id}": {
      "get": {
        "description": "Lekéri egy adott személy adatait az azonosító alapján.",
        "operationId": "SzemelyekController_findID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személy sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin vagy user jogosultság"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy lekérdezése ID alapján",
        "tags": [
          "Személyek"
        ]
      },
      "patch": {
        "description": "Meglévő személy adatainak módosítása admin jogosultsággal.",
        "operationId": "SzemelyekController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSzemelyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Személy sikeresen módosítva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy módosítása",
        "tags": [
          "Személyek"
        ]
      },
      "delete": {
        "description": "Meglévő személy törlése admin jogosultsággal.",
        "operationId": "SzemelyekController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személy sikeresen törölve"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy törlése",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/felhasznalonev/{id}": {
      "get": {
        "description": "Lekéri egy adott személy adatait a felhasználónév alapján.",
        "operationId": "SzemelyekController_findFelhasznaloinev",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személy sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin vagy user jogosultság"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy lekérdezése felhasználónév alapján",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/email/{id}": {
      "get": {
        "description": "Lekéri egy adott személy adatait az email cím alapján.",
        "operationId": "SzemelyekController_findEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személy sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin vagy user jogosultság"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy lekérdezése email alapján",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/profil/me": {
      "get": {
        "description": "Lekéri a bejelentkezett felhasználó adatait.",
        "operationId": "SzemelyekController_findProfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "401": {
            "description": "Nincs bejelentkezett felhasználó vagy érvénytelen JWT token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Saját profil lekérdezése",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/test-validation": {
      "post": {
        "description": "Tesztelési célú végpont adatvalidációhoz (publikus).",
        "operationId": "SzemelyekController_testValidation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Teszt DTO (meghatározatlan struktúra)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validáció sikeres",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          }
        },
        "summary": "Teszt validáció",
        "tags": [
          "Személyek"
        ]
      }
    },
    "/szemelyek/register": {
      "post": {
        "description": "Új személy regisztrálása (publikus végpont).",
        "operationId": "SzemelyekController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSzemelyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Személy sikeresen létrehozva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Szemely"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          }
        },
        "summary": "Személy regisztráció",
        "tags": [
          "Személyek"
        ]
      }
    },
...
    "/feliratkozasok/szemely/{id}": {
      "get": {
        "description": "Lekéri egy adott személyhez tartozó összes eseményt.",
        "operationId": "FeliratkozasokController_szemelyEsemenyei",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Események sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feliratkozas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy eseményeinek lekérdezése",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS keretrendszerrel készített REST API",
    "description": "REST API személyek, címek, események, feliratkozások, hírek és települések kezelésére",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateCimDto": {
        "type": "object",
        "properties": {
          "cim_tipus": {
            "type": "string",
            "description": "A cím típusa",
            "example": "állandó",
            "enum": [
              "állandó",
              "tartózkodási",
              "levelezési"
            ]
          },
          "cim_iranyitoszam": {
            "type": "string",
            "description": "A cím irányítószáma",
            "example": "1234"
          },
          "cim_varos": {
            "type": "string",
            "description": "A cím városa",
            "example": "Budapest"
          },
          "cim_utca": {
            "type": "string",
            "description": "A cím utcája",
            "example": "Fő utca 1"
          }
        },
        "required": [
          "cim_tipus",
          "cim_iranyitoszam",
          "cim_varos",
          "cim_utca"
        ]
      },
...
      "Cim": {
        "type": "object",
        "properties": {
          "cim_id": {
            "type": "number",
            "description": "A cím egyedi azonosítója",
            "example": 1
          },
          "cim_tipus": {
            "type": "string",
            "description": "A cím típusa",
            "example": "állandó",
            "enum": [
              "állandó",
              "tartózkodási",
              "levelezési"
            ]
          },
          "cim_iranyitoszam": {
            "type": "string",
            "description": "A cím irányítószáma",
            "example": "1234"
          },
          "cim_varos": {
            "type": "string",
            "description": "A cím városa",
            "example": "Budapest"
          },
          "cim_utca": {
            "type": "string",
            "description": "A cím utcája",
            "example": "Fő utca 1"
          },
          "cim_szemely_id": {
            "type": "number",
            "description": "A címhez tartozó személy azonosítója",
            "example": 1
          }
        },
        "required": [
          "cim_id",
          "cim_tipus",
          "cim_iranyitoszam",
          "cim_varos",
          "cim_utca",
          "cim_szemely_id"
        ]
      },
      "Esemeny": {
        "type": "object",
        "properties": {
          "esemeny_id": {
            "type": "number",
            "description": "Az esemény egyedi azonosítója",
            "example": 1
          },
          "esemeny_nev": {
            "type": "string",
            "description": "Az esemény neve",
            "example": "Éves konferencia"
          },
          "esemeny_leiras": {
            "type": "string",
            "description": "Az esemény leírása",
            "example": "Éves szakmai konferencia a technológiai újításokról"
          },
          "esemeny_helyszin": {
            "type": "string",
            "description": "Az esemény helyszíne",
            "example": "Budapest, Kongresszusi Központ"
          },
          "esemeny_idopont": {
            "format": "date-time",
            "type": "string",
            "description": "Az esemény időpontja",
            "example": "2025-12-01T10:00:00Z"
          },
          "esemeny_max_fo": {
            "type": "number",
            "description": "Az esemény maximális létszáma",
            "example": 100
          },
          "esemeny_feliratkozasok": {
            "description": "Az eseményhez tartozó feliratkozások",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "esemeny_id",
          "esemeny_nev",
          "esemeny_leiras",
          "esemeny_helyszin",
          "esemeny_idopont",
          "esemeny_max_fo",
          "esemeny_feliratkozasok"
        ]
      },
      "Feliratkozas": {
        "type": "object",
        "properties": {
          "feliratkozas_id": {
            "type": "number",
            "description": "A feliratkozás egyedi azonosítója",
            "example": 1
          },
          "feliratkozas_szemely_id": {
            "type": "number",
            "description": "A feliratkozó személy azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny_id": {
            "type": "number",
            "description": "Az esemény azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny": {
            "description": "A feliratkozás eseménye",
            "allOf": [
              {
                "$ref": "#/components/schemas/Esemeny"
              }
            ]
          }
        },
        "required": [
          "feliratkozas_id",
          "feliratkozas_szemely_id",
          "feliratkozas_esemeny_id",
          "feliratkozas_esemeny"
        ]
      },
      "Szemely": {
        "type": "object",
        "properties": {
          "szemely_id": {
            "type": "number",
            "description": "A személy egyedi azonosítója",
            "example": 1
          },
          "szemely_vezeteknev": {
            "type": "string",
            "description": "A személy vezetékneve",
            "example": "Kovács"
          },
          "szemely_keresztnev": {
            "type": "string",
            "description": "A személy keresztneve",
            "example": "Anna"
          },
          "szemely_szuletesi_datum": {
            "type": "object",
            "description": "A személy születési dátuma (opcionális)",
            "example": "1990-01-01"
          },
          "szemely_anyja_neve": {
            "type": "object",
            "description": "A személy anyjának neve (opcionális)",
            "example": "Szabó Mária"
          },
          "szemely_vegzettseg": {
            "type": "string",
            "description": "A személy végzettsége",
            "example": "középfok",
            "enum": [
              "nincs",
              "alapfok",
              "középfok",
              "felsőfok"
            ]
          },
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe",
            "example": "kovacs.anna@example.com"
          },
          "szemely_telefon": {
            "type": "object",
            "description": "A személy telefonszáma (opcionális)",
            "example": "+36301234567"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava (titkosítva tárolva)",
            "example": "Titkos123"
          },
          "szemely_jogosultsag": {
            "type": "string",
            "description": "A személy jogosultsága",
            "example": "user",
            "enum": [
              "nincs",
              "user",
              "admin"
            ]
          },
          "szemely_torolve": {
            "format": "date-time",
            "type": "string",
            "description": "A személy törlésének időpontja (soft delete)",
            "example": "2025-06-28T13:00:00Z"
          },
          "szemely_cimek": {
            "description": "A személyhez tartozó címek",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cim"
            }
          },
          "szemely_feliratkozasok": {
            "description": "A személyhez tartozó feliratkozások",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feliratkozas"
            }
          }
        },
        "required": [
          "szemely_id",
          "szemely_vezeteknev",
          "szemely_keresztnev",
          "szemely_vegzettseg",
          "szemely_email",
          "szemely_felhasznalonev",
          "szemely_jelszo",
          "szemely_jogosultsag",
          "szemely_cimek",
          "szemely_feliratkozasok"
        ]
      },
...
      "CreateSzemelyDto": {
        "type": "object",
        "properties": {
          "szemely_vezeteknev": {
            "type": "string",
            "description": "A személy vezetékneve",
            "example": "Kovács"
          },
          "szemely_keresztnev": {
            "type": "string",
            "description": "A személy keresztneve",
            "example": "Anna"
          },
          "szemely_szuletesi_datum": {
            "type": "object",
            "description": "A személy születési dátuma (opcionális)",
            "example": "1990-01-01"
          },
          "szemely_anyja_neve": {
            "type": "object",
            "description": "A személy anyjának neve (opcionális)",
            "example": "Szabó Mária"
          },
          "szemely_vegzettseg": {
            "type": "string",
            "description": "A személy végzettsége",
            "example": "középfok",
            "enum": [
              "nincs",
              "alapfok",
              "középfok",
              "felsőfok"
            ]
          },
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe",
            "example": "kovacs.anna@example.com"
          },
          "szemely_telefon": {
            "type": "object",
            "description": "A személy telefonszáma (opcionális)",
            "example": "+36301234567"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava",
            "example": "Titkos123"
          },
          "szemely_jogosultsag": {
            "type": "string",
            "description": "A személy jogosultsága",
            "example": "user",
            "enum": [
              "nincs",
              "user",
              "admin"
            ]
          },
          "szemely_cimek": {
            "description": "A személyhez tartozó címek (opcionális)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCimDto"
            }
          }
        },
        "required": [
          "szemely_vezeteknev",
          "szemely_keresztnev",
          "szemely_vegzettseg",
          "szemely_email",
          "szemely_felhasznalonev",
          "szemely_jelszo",
          "szemely_jogosultsag"
        ]
      },
      "UpdateSzemelyDto": {
        "type": "object",
        "properties": {
          "szemely_vezeteknev": {
            "type": "string",
            "description": "A személy vezetékneve (opcionális)",
            "example": "Kovács"
          },
          "szemely_keresztnev": {
            "type": "string",
            "description": "A személy keresztneve (opcionális)",
            "example": "Anna"
          },
          "szemely_szuletesi_datum": {
            "type": "object",
            "description": "A személy születési dátuma (opcionális)",
            "example": "1990-01-01"
          },
          "szemely_anyja_neve": {
            "type": "object",
            "description": "A személy anyjának neve (opcionális)",
            "example": "Szabó Mária"
          },
          "szemely_vegzettseg": {
            "type": "string",
            "description": "A személy végzettsége (opcionális)",
            "example": "középfok",
            "enum": [
              "nincs",
              "alapfok",
              "középfok",
              "felsőfok"
            ]
          },
          "szemely_email": {
            "type": "string",
            "description": "A személy email címe (opcionális)",
            "example": "kovacs.anna@example.com"
          },
          "szemely_telefon": {
            "type": "object",
            "description": "A személy telefonszáma (opcionális)",
            "example": "+36301234567"
          },
          "szemely_felhasznalonev": {
            "type": "string",
            "description": "A személy felhasználóneve (opcionális)",
            "example": "kovacsanna"
          },
          "szemely_jelszo": {
            "type": "string",
            "description": "A személy jelszava (opcionális)",
            "example": "Titkos123"
          },
          "szemely_jogosultsag": {
            "type": "string",
            "description": "A személy jogosultsága (opcionális)",
            "example": "user",
            "enum": [
              "nincs",
              "user",
              "admin"
            ]
          },
          "szemely_cimek": {
            "description": "A személyhez tartozó címek (opcionális)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCimDto"
            }
          }
        }
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}